upload/requirements.txt
Django==2.2.24
django-stdimage==5.3.0
parso==0.7.0
PyMySQL==1.0.2
pytest==6.2.3
pywin32==302
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
sqlparse==0.4.2
virtualenv==20.10.0
django-crispy-forms